Started working on function for updating the wing kinematics based on the control signal value. Not done yet. Changed the kinematics type to float for the yawff_w_ctlr function. Note, I may want to add a check for the array type after it is passed to the C function from python as yawff and yawff_w_ctlr using different types for this array .

+++ python/libyawff/yawff.py
@@ -299,7 +299,8 @@ def yawff_ctlr_c_wrapper(setpt, config):
     pos = scipy.zeros((n,1), dtype = scipy.dtype('float32'))
     vel = scipy.zeros((n,1), dtype = scipy.dtype('float32'))
     torq = scipy.zeros((n,2), dtype = scipy.dtype('float32'))
-    kine = scipy.zeros((n, config['num_motor']), scipy.dtype('int'))
+    #kine = scipy.zeros((n, config['num_motor']), scipy.dtype('int'))
+    kine = scipy.zeros((n, config['num_motor']), scipy.dtype('float32'))
     u = scipy.zeros((n,1), dtype =scipy.dtype('float32'))
 
     # Create c data structure
